Got tests working so far, github actions

+++ src/libcellgui/models/DiscTypeDistributionTableModel.cpp
@@ -127,7 +127,7 @@ void DiscTypeDistributionTableModel::loadSettings()
     if (discTypeDistribution.empty())
         return;
 
-    beginInsertRows(QModelIndex(), 0, discTypeDistribution.size() - 1);
+    beginInsertRows(QModelIndex(), 0, static_cast<int>(discTypeDistribution.size()) - 1);
     for (const auto& pair : discTypeDistribution)
         rows_.push_back(pair);
     endInsertRows();
